Add initial parts for managing multiple streams
We'll use an array (atomic_t) to track the number of logs for each type. These we only increment, and which we will allocate in RR fashion. We maintain a counter on the total number of logs to know when to start reusing segments instead of allocating new ones for a type (reuse the alloc_type)